Reorganizes and simplifies mapper code
