+++ CONTRIBUTORS.md
@@ -14,3 +14,4 @@ HealthKitOnFHIR contributors
 ====================
 
 * [Paul Schmiedmayer](https://github.com/PSchmiedmayer)
+* [Vishnu Ravi](https://github.com/vishnuravi)
