+++ Sources/HealthKitOnFHIR/Bundle+JSON.swift
@@ -1,0 +1,28 @@
+//
+// This source file is part of the HealthKitOnFHIR open source project
+//
+// SPDX-FileCopyrightText: 2022 Stanford University and the project authors (see CONTRIBUTORS.md)
+//
+// SPDX-License-Identifier: MIT
+//
+
+import Foundation
+
+extension Foundation.Bundle {
+    func decode<T: Decodable>(file: String) -> T {
+        guard let url = self.url(forResource: file, withExtension: nil) else {
+            fatalError("Could not find \(file) in the module.")
+        }
+
+        guard let data = try? Data(contentsOf: url) else {
+            fatalError("Could not load \(file) in the module.")
+        }
+
+        let decoder = JSONDecoder()
+        guard let loadedData = try? decoder.decode(T.self, from: data) else {
+            fatalError("Could not decode \(file) in the module.")
+        }
+
+        return loadedData
+    }
+}
